Move component LinkToInformationAboutEmergencyMeasure from 'components' to 'components/index/SiteTopUpper/WhatsNew'
