+++ components/WhatsNew.vue
@@ -33,7 +33,7 @@
 import Vue from 'vue'
 
 import AppLink from '@/components/AppLink.vue'
-import LinkToInformationAboutEmergencyMeasure from '@/components/LinkToInformationAboutEmergencyMeasure.vue'
+import LinkToInformationAboutEmergencyMeasure from '@/components/index/SiteTopUpper/WhatsNew/LinkToInformationAboutEmergencyMeasure.vue'
 import { convertDateToISO8601Format } from '@/utils/formatDate'
 
 export default Vue.extend({
